requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
defusedxml==0.7.1
filelock==3.18.0
Flask==3.1.1
Flask-Dropzone==2.0.0
fonttools==4.58.4
fpdf==1.7.2
fpdf2==2.8.3
fsspec==2025.5.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
gunicorn==23.0.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
ninja==1.11.1.4
numpy==2.3.0
packaging==25.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
python-bidi==0.6.6
PyYAML==6.0.2
reportlab==4.4.2
requests==2.32.4
rsa==4.9.1
scikit-image==0.25.2
scipy==1.15.3
setuptools==80.9.0
shapely==2.1.1
sympy==1.14.0
tifffile==2025.6.11
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
